[future: Rivsen/wand#1]: split code into lib
